market_analysis/screeners/requirements.txt
nsepy
pandas
requests
futures
datetime
config
python-decouple
#pandas.core.tools.datetimes
